requirements.txt

langchain
langchain_community

pinecone-client
openai
langchain

google-generativeai
langchain-openai
streamlit
 


